+++ package.json
@@ -29,7 +29,8 @@
     "react-scripts": "5.0.1",
     "react-toastify": "^11.0.5",
     "swiper": "^11.1.9",
-    "web-vitals": "^2.1.4"
+    "web-vitals": "^2.1.4",
+    "zustand": "^5.0.5"
   },
   "scripts": {
     "start": "react-scripts start",
